requirements.txt
pyTelegramBotAPI==4.1.1
requests==2.26.0
python-decouple==3.5
python-telegram-bot-calendar==1.0.5